Simplifying the ThreadLocal context API
