+++ core/azure-core/src/main/java/com/azure/core/util/Context.java
@@ -3,6 +3,8 @@
 
 package com.azure.core.util;
 
+import java.lang.reflect.Array;
+import java.lang.reflect.Field;
 import java.util.Optional;
 
 /**
@@ -11,7 +13,7 @@ import java.util.Optional;
  * {@code null}. Each context object is immutable. The {@code addData(Object, Object)} method creates a new
  * {@code Context} object that refers to its parent, forming a linked list.
  */
-public class Context {
+public class Context implements AutoCloseable {
     // All fields must be immutable.
     //
     /**
@@ -19,9 +21,14 @@ public class Context {
      */
     public static final Context NONE = new Context(null, null, null);
 
-    final Context parent;
-    final Object key;
-    final Object value;
+    private final Context parent;
+    private final Object key;
+    private final Object value;
+
+    private ThreadLocal<Context> threadLocal;
+    private static boolean reflectionInitiated = false;
+    private static Field threadLocalsField;
+    private static Field tableField;
 
     /**
      * Constructs a new {@link Context} object.
@@ -30,18 +37,22 @@ public class Context {
      * @param value the value
      */
     public Context(Object key, Object value) {
-        if (key == null) {
-            throw new IllegalArgumentException("key cannot be null");
-        }
-        this.parent = null;
-        this.key = key;
-        this.value = value;
+        this(null, key, value);
     }
 
     Context(Context parent, Object key, Object value) {
+        this(parent, key, value, new ThreadLocal<>());
+    }
+
+    Context(Context parent, Object key, Object value, ThreadLocal<Context> threadLocal) {
         this.parent = parent;
         this.key = key;
         this.value = value;
+        this.threadLocal = threadLocal;
+
+        if (threadLocal != null && key != null && value != null) {
+            threadLocal.set(this);
+        }
     }
 
     /**
@@ -56,7 +67,14 @@ public class Context {
         if (key == null) {
             throw new IllegalArgumentException("key cannot be null");
         }
-        return new Context(this, key, value);
+
+        // when we add data, the new ThreadLocalContext becomes the root element,
+        // and so we remove this context out of thread local
+        if (threadLocal != null) {
+            threadLocal.remove();
+        }
+
+        return new Context(this, key, value, threadLocal);
     }
 
     /**
@@ -78,4 +96,61 @@ public class Context {
         return Optional.empty();
     }
 
+    @Override
+    public void close() {
+        // Because of how we use context, we can not simply call threadLocal.remove() here, as
+        // the context created at the start of a try-with-resources will not necessarily be the one
+        // that needs to have close() called on it, and from that context we have no means on finding
+        // the right context, except for diving into the thread reflectively.
+        Context ctx = getThreadLocalContext();
+        if (ctx != null && ctx.threadLocal != null) {
+            ctx.threadLocal.remove();
+        }
+    }
+
+    // FIXME temporary toString for debugging - delete before merging
+    @Override
+    public String toString() {
+        if (key == null && value == null) {
+            return "[Context NONE, parent=" + parent + "]";
+        }
+        return "[Context key='" + key + "', value='" + value + "', parent=" + parent + "]";
+    }
+
+    public static Context getThreadLocalContext() {
+        Context result = Context.NONE;
+
+        try {
+            if (!reflectionInitiated) {
+                threadLocalsField = Thread.class.getDeclaredField("threadLocals");
+                threadLocalsField.setAccessible(true);
+
+                Class threadLocalMapClass = Class.forName("java.lang.ThreadLocal$ThreadLocalMap");
+                tableField = threadLocalMapClass.getDeclaredField("table");
+                tableField.setAccessible(true);
+
+                reflectionInitiated = true;
+            }
+
+            Object table = tableField.get(threadLocalsField.get(Thread.currentThread()));
+
+            for (int i = 0, max = Array.getLength(table); i < max; i++) {
+                Object entry = Array.get(table, i);
+                if (entry != null) {
+                    Field valueField = entry.getClass().getDeclaredField("value");
+                    valueField.setAccessible(true);
+                    Object value = valueField.get(entry);
+
+                    if (value instanceof Context) {
+                        result = (Context) value;
+                        break;
+                    }
+                }
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return result;
+    }
 }
